+++ packages/api/eslint-config-nestjs/custom-rules/plugin.js
@@ -1,0 +1,16 @@
+import { eslintCompatPlugin } from '@oxlint/plugins'
+
+import checkApiPropertyTypes from './check-api-property-types.js'
+import checkColumnTypes from './check-column-types.js'
+import checkMissingTranslations from './check-missing-translations.js'
+
+export default eslintCompatPlugin({
+  meta: {
+    name: 'custom-rules'
+  },
+  rules: {
+    'check-api-property-types': checkApiPropertyTypes,
+    'check-column-types': checkColumnTypes,
+    'check-missing-translations': checkMissingTranslations
+  }
+})