+++ packages/api/eslint-config-nestjs/test/fixtures/check-api-property-types/support/uuid.ts
@@ -1,3 +1,8 @@
-export type Uuid<Brand extends string> = string
-  & { readonly __brand: 'uuid' }
-  & { readonly __uuid: Brand }
+/* eslint-disable @typescript-eslint/naming-convention */
+export type Uuid<Brand extends string> = string & 
+(
+  ({ readonly __brand: 'uuid' }
+    & { readonly __uuid: Brand }) 
+| ({ readonly _brand: 'uuid' } 
+  & { readonly _uuid: Brand })
+)